Correciones del control de gasto y agregacion de nueva logica en el historial
