+++ src/pages/historial/index.tsx
@@ -1,47 +1,75 @@
-import React from 'react';
-import { Card } from 'antd';
-import { useLocation } from 'react-router-dom';
+import { useState, useEffect } from 'react';
+import axios from 'axios';
+import { Table, Tag, message } from 'antd';
+import moment from 'moment';
 
-const Historial = () => {
-  const location = useLocation();
-  const gastos = location.state?.gastos || [];
-  const saldoRestante = location.state?.saldoRestante || 0;
-  const gastoEmergencia = location.state?.gastoEmergencia || 0; // Agrega esta línea para obtener el gasto de emergencia
+interface HistorialItem {
+  id: number;
+  name: string;
+  amount: number;
+  date: string;
+  type: string; // 'Gasto Fijo', 'Gasto Variable' o 'Ahorro'
+}
+
+export function Historial() {
+  const [historial, setHistorial] = useState<HistorialItem[]>([]);
+
+  useEffect(() => {
+    fetchHistorial();
+  }, []);
+
+  const fetchHistorial = async () => {
+    try {
+      const response = await axios.get('http://localhost:4000/historial');
+      const data = response.data.map((item: HistorialItem, index: number) => ({
+        ...item,
+        uniqueId: `${item.type}-${item.id}-${index}` // Crear un id único
+      }));
+      setHistorial(data);
+    } catch (error) {
+      console.error('Error fetching historial:', error);
+      message.error('Error al obtener el historial.'); // Muestra un mensaje de error al usuario
+    }
+  };
+
+  // Columnas de la tabla
+  const columns = [
+    {
+      title: 'Nombre',
+      dataIndex: 'name',
+      key: 'name',
+    },
+    {
+      title: 'Monto',
+      dataIndex: 'amount',
+      key: 'amount',
+      render: (amount: number) => {
+        if (typeof amount !== 'number' || isNaN(amount)) {
+          return 'N/A';
+        }
+        return `$ ${amount.toFixed(2)}`;
+      },
+    },
+    {
+      title: 'Fecha',
+      dataIndex: 'date',
+      key: 'date',
+      render: (date: string) => moment(date).format('DD/MM/YYYY'), 
+    },
+    {
+      title: 'Tipo',
+      dataIndex: 'type',
+      key: 'type',
+      render: (type: string) => (
+        <Tag color={type === 'Gasto Fijo' ? 'red' : type === 'Gasto Variable' ? 'orange' : 'blue'}>{type}</Tag>
+      ),
+    },
+  ];
 
   return (
-    <div className="container mt-5 text-white">
-      <h1>Historial de Gastos</h1>
-      <Card>
-        {gastos.length === 0 ? (
-          <p>No hay gastos registrados.</p>
-        ) : (
-          <>
-            <p><strong>Saldo Restante:</strong> ${saldoRestante}</p>
-            <p><strong>Gasto de Emergencia:</strong> ${gastoEmergencia}</p> {/* Agrega esta línea para mostrar el gasto de emergencia */}
-            {gastos.map((gasto: { nombre: string | number | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | null | undefined; cantidad: string | number | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | null | undefined; tipo: string | number | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | null | undefined; fecha: string | number | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | null | undefined; }, index: React.Key | null | undefined) => (
-              <div key={index}>
-                <h4>{gasto.nombre}</h4>
-                <p>
-                  <strong>Cantidad:</strong> ${gasto.cantidad}
-                </p>
-                <p>
-                  <strong>Tipo:</strong> {gasto.tipo}
-                </p>
-                <p>
-                  <strong>Fecha:</strong> {gasto.fecha}
-                </p>
-                <p>
-                  <strong>Fecha:</strong> {gasto.fecha}
-                </p>
-                
-                <hr />
-              </div>
-            ))}
-          </>
-        )}
-      </Card>
+    <div style={{ backgroundColor: '#fff', padding: '20px', borderRadius: '8px', boxShadow: '0 2px 6px rgba(0, 0, 0, 0.1)' }}>
+      <h2>Historial</h2>
+      <Table columns={columns} dataSource={historial} rowKey="uniqueId" />
     </div>
   );
-};
-
-export default Historial;
+}
